test(ai): rename domain_planners test files to match split source files (Refs #2289)
Aligns AI test file names with the now-split source files under
`packages/colonizethis_ai/lib/src/planning/`:

- `domain_planners_test_part1_test.dart` -> `domain_planner_orchestrator_test.dart`
- `domain_planners_test_part2_test.dart` -> `war_desire_calculator_test.dart`
- `domain_planners_test_part3_test.dart` -> `diplomatic_candidate_scoring_test.dart`
- `domain_planners_civilian_work_log_test.dart` -> `domain_planner_orchestrator_civilian_work_log_test.dart`
- `domain_planners_run_move_work_xor_test.dart` -> `domain_planner_orchestrator_move_work_xor_test.dart`
- `domain_planners_test_fake_api.dart` -> `domain_planner_test_fake_api.dart`

All renames use `git mv` so blame/history is preserved. The four tests
that import the shared fake-suggestion-API helper update their relative
import path from `domain_planners_test_fake_api.dart` to
`domain_planner_test_fake_api.dart`. No assertion or test body changes.
