+++ packages/colonizethis_ai/test/domain_planner_orchestrator_civilian_work_log_test.dart
@@ -5,7 +5,7 @@ import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:logger/logger.dart';
 
-import 'domain_planners_test_fake_api.dart';
+import 'domain_planner_test_fake_api.dart';
 
 void main() {
   group('runDomainPlanners civilian work logging', () {
